Move method to get GQL event query result to util
