+++ packages/codegen/src/templates/hooks-template.handlebars
@@ -4,9 +4,9 @@
 
 import assert from 'assert';
 
-// import { updateStateForMappingType, updateStateForElementaryType } from '@cerc-io/util';
+import { updateStateForMappingType, updateStateForElementaryType, ResultEvent } from '@cerc-io/util';
 
-import { Indexer, ResultEvent } from './indexer';
+import { Indexer } from './indexer';
 
 /**
  * Hook function to store an initial state.
